tools/python/requirements.txt
python-chess>=1.999
pyserial>=3.5
